requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
PyJWT==2.10.1
python-dotenv==1.0.1
werkzeug==3.1.3
google-genai==1.10.0
